refactor: 完成项目全面升级与功能完善
整体变更包括：
1.  统一将浮点金额类型替换为BigDecimal，修复精度问题
2.  完善JWT认证与配置，添加密钥校验，优化请求头处理
3.  添加RBAC权限系统与管理员功能，包括商品审核、用户角色管理
4.  实现购物车本地持久化
5.  清理调试日志，优化服务层逻辑
6.  更新文档与配置文件

+++ server/src/main/java/me/code/springboot_postgres/dtos/responses/entities/FavoriteDTO.java
@@ -1,10 +1,12 @@
 package me.code.springboot_postgres.dtos.responses.entities;
 
+import java.math.BigDecimal;
+
 public record FavoriteDTO(
         String id,
         String productId,
         String productName,
         String productImageUrl,
-        double productPrice,
+        BigDecimal productPrice,
         String createdAt) {
 }
